% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_Single_Variants.R
\name{Annotate_Single_Variants}
\alias{Annotate_Single_Variants}
\title{Functionally annotate a list of variants}
\usage{
Annotate_Single_Variants(
  agds_dir,
  single_variants_list,
  QC_label = "annotation/filter",
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Annotation_name
)
}
\arguments{
\item{agds_dir}{file directory of annotated GDS (aGDS) files for all chromosomes (1-22).}

\item{single_variants_list}{a data frame containing the information of variants to be functionally annotated. The data frame must include 4 columns with
the following names: "CHR" (chromosome number), "POS" (position), "REF" (reference allele), and "ALT" (alternative allele).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file  (default = "annotation/filter").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file \cr (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the annotation names and the corresponding channel names in the aGDS file.}

\item{Annotation_name}{a vector of qualitative/quantitative annotation names user wants to extract.}
}
\value{
A data frame containing the basic information (chromosome, position, reference allele and alternative allele)
and annotation scores for the input variants.
}
\description{
The \code{Annotate_Single_Variants} function takes in a list of variants to functionally annotate the input variants
}
\references{
Li, Z., Li, X., et al. (2022). A framework for detecting
noncoding rare-variant associations of large-scale whole-genome sequencing
studies. \emph{Nature Methods}, \emph{19}(12), 1599-1611.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})
}
