% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nullmodel_genome_cond_spa.R
\name{fit_nullmodel_genome_cond_spa}
\alias{fit_nullmodel_genome_cond_spa}
\title{Fitting conditional generalized linear mixed models with known relationship matrices
for conditional analysis in imbalanced case-control setting.}
\usage{
fit_nullmodel_genome_cond_spa(
  fixed,
  data = parent.frame(),
  kins,
  use_sparse = TRUE,
  use_SPA = TRUE,
  agds_dir,
  known_loci,
  geno_missing_imputation = c("mean", "minor"),
  MAC_cutoff = 20,
  output_path,
  cond_null_model_name = NULL,
  phenotype_id,
  phenotype,
  kins_cutoff = 0.022,
  id,
  random.slope = NULL,
  groups = NULL,
  family = binomial(link = "logit"),
  method = "REML",
  method.optim = "AI",
  maxiter = 500,
  tol = 1e-05,
  taumin = 1e-05,
  taumax = 1e+05,
  tauregion = 10,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fixed}{an object of class \code{\link{formula}} (or one that can be coerced to that class):
a symbolic description of the fixed effects model to be fitted. For multiple phenotype analysis,
\code{\link{formula}} recognized by \code{\link{lm}}, such as \code{cbind(y1,y2,y3) ~ x1 + x2},
can be used in \code{fixed} as fixed effects.}

\item{data}{a data frame or list (or object coercible by \code{as.data.frame} to a data frame)
containing the variables in the model.}

\item{kins}{a known positive semi-definite relationship matrix
(e.g. kinship matrix in genetic association studies) or a list of known
positive semi-definite relationship matrices. The rownames and colnames of
these matrices must at least include all samples as specified in the \code{id} column
of the data frame \code{data}. If \code{kins} is NULL, \code{fit_nullmodel}
will switch to the generalized linear model with no random effects.}

\item{use_sparse}{a logical switch of whether the provided dense \code{kins} matrix should be
transformed to a sparse matrix (default = TRUE).}

\item{use_SPA}{a logical switch determines if the null model fitting occurs in an imbalanced case-control setting (default = TRUE).}

\item{agds_dir}{file directory of annotated GDS (aGDS) files for all chromosomes (1-22)}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT)  (default = NULL).}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{MAC_cutoff}{the cutoff of the minimum minor allele count of known variants adjusted in
conditional analysis (default = 20).}

\item{output_path}{the directory for the output files.}

\item{cond_null_model_name}{the file name of conditional null models (default = NULL).}

\item{phenotype_id}{id of samples.}

\item{phenotype}{outcome in regression.}

\item{kins_cutoff}{the cutoff value for clustering samples to make the output matrix sparse block-diagonal
(default = 0.022).}

\item{id}{a column in the data frame \code{data}, indicating the id of samples.
When there are duplicates in \code{id}, the data is assumed to be longitudinal with repeated measures.}

\item{random.slope}{an optional column indicating the random slope for time effect used
in a mixed effects model for longitudinal data. It must be included in the names of \code{data}.
There must be duplicates in \code{id} and \code{method.optim} must be "AI" (default = NULL).}

\item{groups}{an optional categorical variable indicating the groups used in a
heteroscedastic linear mixed model (allowing residual variances in different groups to be different).
This variable must be included in the names of \code{data}, and \code{family} must be "gaussian"
and \code{method.optim} must be "AI" (default = NULL).}

\item{family}{a description of the error distribution and link function to be used
in the model. This can be a character string naming a family function, a family
function or the result of a call to a family function. (See \code{\link{family}} for details of family functions).}

\item{method}{method of fitting the generalized linear mixed model. Either "REML" or "ML" (default = "REML").}

\item{method.optim}{optimization method of fitting the generalized linear mixed model.
Either "AI", "Brent" or "Nelder-Mead" (default = "AI").}

\item{maxiter}{a positive integer specifying the maximum number of iterations when
fitting the generalized linear mixed model (default = 500).}

\item{tol}{a positive number specifying tolerance, the difference threshold for parameter
estimates below which iterations should be stopped (default = 1e-5).}

\item{taumin}{the lower bound of search space for the variance component parameter \eqn{\tau} (default = 1e-5),
used when \code{method.optim} = "Brent". See Details.}

\item{taumax}{the upper bound of search space for the variance component parameter \eqn{\tau} (default = 1e5),
used when \code{method.optim} = "Brent". See Details.}

\item{tauregion}{the number of search intervals for the REML or ML estimate of the variance component
parameter \eqn{\tau} (default = 10), used when \code{method.optim} = "Brent". See Details.}

\item{verbose}{a logical switch for printing detailed information (parameter estimates in each iteration)
for testing and debugging purpose (default = FALSE).}

\item{...}{additional arguments that could be passed to \code{\link{glm}}.}
}
\value{
The function returns objects of the null models fit from \code{\link{fit_nullmodel}} 
and whether the \code{kins} matrix is sparse when fitting the null model, each chromosome has one output. 
See \code{\link{fit_nullmodel}} for more details.
}
\description{
The \code{fit_nullmodel_genome_cond_spa} function fit regression models for conditional analysis in imbalanced 
case-control setting, which provides the preliminary step for subsequent conditional variant-set tests in
conditional analysis. Each chromosome has a separate null model for conditional analysis. See \code{fit_nullmodel} for more details.
}
\references{
Chen, H., et al. (2016). Control for population structure and relatedness for binary traits
in genetic association studies via logistic mixed models. \emph{The American Journal of Human Genetics}, \emph{98}(4), 653-666.
(\href{https://doi.org/10.1016/j.ajhg.2016.02.012}{pub})

Chen, H., et al. (2019). Efficient variant set mixed model association tests for continuous and
binary traits in large-scale whole-genome sequencing studies. \emph{The American Journal of Human Genetics}, \emph{104}(2), 260-274.
(\href{https://doi.org/10.1016/j.ajhg.2018.12.012}{pub})

Chen, H. (2021). GMMAT: Generalized linear Mixed Model Association Tests Version 1.3.2.
(\href{https://cloud.r-project.org/web/packages/GMMAT/vignettes/GMMAT.pdf}{web})
}
