% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Centric_Coding_Results_Summary.R
\name{Gene_Centric_Coding_Results_Summary}
\alias{Gene_Centric_Coding_Results_Summary}
\title{Summarize gene-centric coding analysis results generated by \code{STAARpipeline} package and
perform conditional analysis for (unconditionally) significant coding masks by adjusting for a given list of known variants.}
\usage{
Gene_Centric_Coding_Results_Summary(
  agds_dir,
  gene_centric_coding_jobs_num,
  input_path,
  output_path,
  gene_centric_results_name,
  obj_nullmodel,
  known_loci = NULL,
  method_cond = c("optimal", "naive"),
  QC_label = "annotation/filter",
  geno_missing_imputation = c("mean", "minor"),
  variant_type = c("SNV", "Indel", "variant"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = FALSE,
  Annotation_name = NULL,
  alpha = 2.5e-06,
  manhattan_plot = FALSE,
  QQ_plot = FALSE
)
}
\arguments{
\item{agds_dir}{file directory of annotated GDS (aGDS) files for all chromosomes (1-22)}

\item{gene_centric_coding_jobs_num}{the number of gene-centric coding analysis results generated by \code{STAARpipeline} package.}

\item{input_path}{the directory of gene-centric coding analysis results that generated by \code{STAARpipeline} package.}

\item{output_path}{the directory for the output files.}

\item{gene_centric_results_name}{file name of gene-centric coding analysis results generated by \code{STAARpipeline} package.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{fit_nullmodel} function in the \code{STAARpipeline} package,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{genesis2staar_nullmodel} function in the \code{STAARpipeline} package.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT).}

\item{method_cond}{a character value indicating the method for conditional analysis.
\code{optimal} refers to regressing residuals from the null model on \code{known_loci}
as well as all covariates used in fitting the null model (fully adjusted) and taking the residuals;
\code{naive} refers to regressing residuals from the null model on \code{known_loci}
and taking the residuals (default = \code{optimal}).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file  (default = "annotation/filter").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{variant_type}{variants include in the analysis. Choices include "variant", "SNV", or "Indel" (default = "SNV").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{Use_annotation_weights}{use annotations as weights or not (default = FALSE).}

\item{Annotation_name}{annotations used in STAAR.}

\item{alpha}{p-value threshold of significant results (default=2.5E-06).}

\item{manhattan_plot}{output manhattan plot or not (default = FALSE).}

\item{QQ_plot}{output Q-Q plot or not (default = FALSE).}
}
\value{
The function returns the following analysis results:

\code{coding_sig.csv}: a matrix that summarizes the unconditional significant coding masks detected by STAAR-O (STAAR-O pvalue smaller than the threshold alpha),
including gene name ("Gene name"), chromosome ("chr"), coding functional category ("Category"), number of variants  ("#SNV"),
and unconditional p-values of set-based tests SKAT ("SKAT(1,25)"), Burden ("Burden(1,1)"), ACAT-V ("ACAT-V(1,25)") and STAAR-O ("STAAR-O").

\code{coding_sig_cond.csv}: a matrix that summarized the conditional analysis results of unconditional significant coding masks detected by STAAR-O (available if known_loci is not a NULL),
including gene name ("Gene name"), chromosome ("chr"), coding functional category ("Category"), number of variants  ("#SNV"),
and conditional p-values of set-based tests SKAT ("SKAT(1,25)"), Burden ("Burden(1,1)"), ACAT-V ("ACAT-V(1,25)") and STAAR-O ("STAAR-O").

\code{results_plof_genome.Rdata}: a matrix contains the STAAR p-values (including STAAR-O) of the coding mask defined by the putative loss of function variants (plof) for all protein-coding genes across the genome.

\code{plof_sig.csv}: a matrix contains the unconditional STAAR p-values (including STAAR-O) of the unconditional significant plof masks.

\code{plof_sig_cond.csv}: a matrix contains the conditional STAAR p-values (including STAAR-O) of the unconditional significant plof masks (available if known_loci is not a NULL).

\code{results_plof_ds_genome.Rdata}: a matrix contains the STAAR p-values (including STAAR-O) of the coding mask defined by the putative loss of function variants and disruptive missense variants (plof_ds) for all protein-coding genes across the genome.

\code{plof_ds_sig.csv}: a matrix contains the unconditional STAAR p-values (including STAAR-O) of the unconditional significant plof_ds masks.

\code{plof_ds_sig_cond.csv}: a matrix contains the conditional STAAR p-values (including STAAR-O) of the unconditional significant plof_ds masks (available if known_loci is not a NULL).

\code{results_disruptive_missense_genome.Rdata}: a matrix contains the STAAR p-values (including STAAR-O) of the coding mask defined by the disruptive missense variants (disruptive_missense) for all protein-coding genes across the genome.

\code{disruptive_missense_sig.csv}: a matrix contains the unconditional STAAR p-values (including STAAR-O) of the unconditional significant disruptive_missense masks.

\code{disruptive_missense_sig_cond.csv}: a matrix contains the conditional STAAR p-values (including STAAR-O) of the unconditional significant disruptive_missense masks (available if known_loci is not a NULL).

\code{results_missense_genome.Rdata}: a matrix contains the STAAR p-values (including STAAR-O) of the coding mask defined by the missense variants (missense) for all protein-coding genes across the genome.

\code{missense_sig.csv}: a matrix contains the unconditional STAAR p-values (including STAAR-O) of the unconditional significant missense masks.

\code{missense_sig_cond.csv}: a matrix contains the conditional STAAR p-values (including STAAR-O) of the unconditional significant missense masks (available if known_loci is not a NULL).

\code{results_synonymous_genome.Rdata}: a matrix contains the STAAR p-values (including STAAR-O) of the coding mask defined by the synonymous variants (synonymous) for all protein-coding genes across the genome.

\code{synonymous_sig.csv}: a matrix contains the unconditional STAAR p-values (including STAAR-O) of the unconditional significant synonymous masks.

\code{synonymous_sig_cond.csv}: a matrix contains the conditional STAAR p-values (including STAAR-O) of the unconditional significant synonymous masks (available if known_loci is not a NULL).
}
\description{
The \code{Gene_Centric_Coding_Results_Summary} function takes in the objects of gene-centric coding analysis results
generated by \code{STAARpipeline} package,
the object from fitting the null model, and the set of known variants to be adjusted for in conditional analysis
to summarize the gene-centric coding analysis results and analyze the conditional association between a quantitative/dichotomous phenotype and
the rare variants in the unconditional significant coding masks.
}
