% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Analysis_Results_Summary.R
\name{Individual_Analysis_Results_Summary}
\alias{Individual_Analysis_Results_Summary}
\title{Summarize individual-variant analysis results generated by \code{STAARpipeline} package}
\usage{
Individual_Analysis_Results_Summary(
  agds_dir,
  jobs_num,
  input_path,
  output_path,
  individual_results_name,
  obj_nullmodel,
  known_loci = NULL,
  method_cond = c("optimal", "naive"),
  QC_label = "annotation/filter",
  geno_missing_imputation = c("mean", "minor"),
  alpha = 5e-09,
  manhattan_plot = FALSE,
  QQ_plot = FALSE
)
}
\arguments{
\item{agds_dir}{a data farme containing directory of GDS/aGDS files.}

\item{jobs_num}{a data frame containing the number of analysis results, including the number of individual analysis results,
the number of sliding window analysis results, and the number of dynamic window analysis results.}

\item{input_path}{the directory of individual analysis results that generated by \code{STAARpipeline} package.}

\item{output_path}{the directory for the output files.}

\item{individual_results_name}{the file name of individual analysis results generated by \code{STAARpipeline} package.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{fit_nullmodel} function in the \code{STAARpipeline} package,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{genesis2staar_nullmodel} function in the \code{STAARpipeline} package.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT) (default = NULL).}

\item{method_cond}{a character value indicating the method for conditional analysis.
\code{optimal} refers to regressing residuals from the null model on \code{known_loci}
as well as all covariates used in fitting the null model (fully adjusted) and taking the residuals;
\code{naive} refers to regressing residuals from the null model on \code{known_loci}
and taking the residuals (default = \code{optimal}).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file.}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{alpha}{p-value threshold of significant results (default=5E-09).}

\item{manhattan_plot}{output manhattan plot or not (default = FALSE).}

\item{QQ_plot}{output Q-Q plot or not (default = FALSE).}
}
\value{
The function returns the following analysis results:

\code{results_individual_analysis_genome.Rdata}: a matrix contains the score test p-value and effect size estimation of each variant across the genome.

\code{results_individual_analysis_sig.Rdata} and \code{results_individual_analysis_sig.csv}: a matrix contains the score test p-values and effect size estimations of significant results (p-value < alpha).

\code{results_sig_cond.Rdata} and \code{results_sig_cond.csv}: a matrix contains the conditional score test p-values for each significant variant  (available if known_loci is not a NULL).
}
\description{
The \code{Individual_Analysis_Results_Summary} function takes in the objects of individual analysis results
generated by \code{STAARpipeline} package,
the object from fitting the null model, and the set of known variants to be adjusted for in conditional analysis
to summarize the individual analysis results and analyze the conditional association between a quantitative/dichotomous phenotype and
the unconditional significant single variants.
}
