% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Variants_List_Analysis.R
\name{Single_Variants_List_Analysis}
\alias{Single_Variants_List_Analysis}
\title{Calculate individual-variant p-values of a list of variants}
\usage{
Single_Variants_List_Analysis(
  agds_dir,
  single_variants_list,
  obj_nullmodel,
  QC_label = "annotation/filter",
  geno_missing_imputation = c("mean", "minor")
)
}
\arguments{
\item{agds_dir}{file directory of annotated GDS (aGDS) files for all chromosomes (1-22).}

\item{single_variants_list}{name a data frame containing the information of variants to be functionally annotated. The data frame must include 4 columns with
the following names: "CHR" (chromosome number), "POS" (position), "REF" (reference allele), and "ALT" (alternative allele).}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{fit_nullmodel} function in the \code{STAARpipeline} package,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{genesis2staar_nullmodel} function in the \code{STAARpipeline} package.}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}
}
\value{
a data frame containing the basic information (chromosome, position, reference allele and alternative allele)
the score test p-values, and the effect sizes for the input variants.
}
\description{
The \code{Single_Variants_List_Analysis} function takes in a list of variants to calculate the p-values and effect sizes of the input variants
}
